requirements.txt
Flask==1.1.2
python-dotenv==0.13.0
black==19.10b0
jedi==0.17.1
pylint==2.5.3
redis==3.5.3
rq==1.4.2
